+++ app/src/main/java/eu/huberisation/moncarnetcovid/data/migrations/Migration0to1.kt
@@ -1,0 +1,4 @@
+package eu.huberisation.moncarnetcovid.data.migrations
+
+object Migration_0_to_1 {
+}