+++ app/src/main/java/eu/huberisation/moncarnetcovid/decoder/CborDecoder.kt
@@ -1,0 +1,4 @@
+package eu.huberisation.moncarnetcovid.decoder
+
+class CborDecoderr {
+}