+++ app/src/main/java/eu/huberisation/moncarnetcovid/decoder/CoseDecoder.kt
@@ -1,0 +1,7 @@
+package eu.huberisation.moncarnetcovid.decoder
+
+object CoseService {
+    fun decode(input: ByteArray) {
+
+    }
+}