+++ app/src/main/java/eu/huberisation/moncarnetcovid/entities/CertificatFactory.kt
@@ -1,0 +1,4 @@
+package eu.huberisation.moncarnetcovid.entities
+
+class CertificatFactory {
+}