+++ app/src/main/java/eu/huberisation/moncarnetcovid/ui/onboarding/OnboardingActivity.kt
@@ -1,0 +1,15 @@
+package eu.huberisation.moncarnetcovid.ui.onboarding
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import android.view.View
+import android.view.Window
+import androidx.navigation.findNavController
+import eu.huberisation.moncarnetcovid.R
+
+class OnboardingActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_onboarding)
+    }
+}