+++ app/src/main/java/eu/huberisation/moncarnetcovid/viewmodel/MainActivityViewModel.kt
@@ -1,0 +1,4 @@
+package eu.huberisation.moncarnetcovid.ui
+
+class MainActivityViewModel {
+}